+++ examples/angular/sample-app/src/app/app.component.ts
@@ -7,6 +7,7 @@ import { RudderAnalytics } from '@rudderstack/analytics-js';
   selector: 'app-root',
   templateUrl: './app.component.html',
   styleUrls: ['./app.component.css'],
+  standalone: false,
 })
 export class AppComponent implements OnInit, OnDestroy {
   title = 'sample-app';