+++ examples/integrations/Ninetailed/sample-apps/app-using-v3-cdn/src/config.js
@@ -1,8 +1,7 @@
-const DATAPLANE_URL = 'http://localhost:8080';
-const WRITE_KEY = 'dummyWriteKey';
-const CONFIG_URL = 'https://api.dev.rudderlabs.com';
+const DATAPLANE_URL = '<dataPlaneUrl>';
+const WRITE_KEY = '<writeKey>';
+
 module.exports = {
   DATAPLANE_URL,
   WRITE_KEY,
-  CONFIG_URL,
 };