+++ packages/analytics-js-common/src/constants/integrations/PostAffiliatePro/constants.ts
@@ -1,6 +1,9 @@
+import {
+  POST_AFFILIATE_PRO_NAME as NAME,
+  POST_AFFILIATE_PRO_DISPLAY_NAME as DISPLAY_NAME,
+} from '../Destinations';
+
 const DIR_NAME = 'PostAffiliatePro';
-const NAME = 'POST_AFFILIATE_PRO';
-const DISPLAY_NAME = 'Post Affiliate Pro';
 
 const DISPLAY_NAME_TO_DIR_NAME_MAP = { [DISPLAY_NAME]: DIR_NAME };
 const CNameMapping = {