+++ packages/analytics-js-common/src/constants/integrations/RollBar/constants.ts
@@ -1,6 +1,6 @@
+import { ROLLBAR_NAME as NAME, ROLLBAR_DISPLAY_NAME as DISPLAY_NAME } from '../Destinations';
+
 const DIR_NAME = 'RollBar';
-const NAME = 'ROLLBAR';
-const DISPLAY_NAME = 'RollBar';
 
 const DISPLAY_NAME_TO_DIR_NAME_MAP = { [DISPLAY_NAME]: DIR_NAME };
 const CNameMapping = {