+++ packages/analytics-js-integrations/src/integrations/GA4_V2/browser.js
@@ -34,7 +34,7 @@ export default class GA4_V2 extends GA4 {
     if (!isDefinedAndNotNull(newConfig.measurementId)) {
       logger.error('Measurement ID is required for GA4');
     }
-    
+
     newConfig.isExtendedGa4_V2 = true;
     super(newConfig, analytics, destinationInfo);
     this.analytics = analytics;